새로운 방 생성 요청 구현
