+++ views/lobby.ejs
@@ -58,7 +58,7 @@
 <div id="create-game-modal" class="modal fade" role="dialog">
     <div class="modal-dialog">
         <div class="modal-content">
-            <form class="form-horizontal" action="/create" method="POST">
+            <form action="/game/create" method="POST" class="form-horizontal">
                 <div class="modal-header">
                     <button type="button" class="close" data-dismiss="modal">&times;</button>
                     <h4 class="modal-title text-center">방 만들기</h4>
@@ -67,8 +67,8 @@
                     <div class="form-group">
                         <label for="room-name" class="col-sm-2 control-label">제목</label>
                         <div class="col-sm-10">
-                            <input type="text" name="room_name" class="form-control" id="room-name" placeholder="방 제목"
-                                   autofocus>
+                            <input type="text" class="form-control" id="room-name" name="roomName" placeholder="방 제목"
+                                   required autofocus>
                         </div>
                     </div>
                 </div>
